requirements.txt
Appium-Python-Client~=2.2.0
selenium~=4.1.3